Property/src/test/java/PropertyTests/CommuneTest.java
package PropertyTests;

/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

import com.mycompany.property.app.appConfiguration;
import com.mycompany.property.model.propertyPackage.Commune;
import junit.framework.Assert;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.AnnotationConfigApplicationContext;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author donkey
 */
public class CommuneTest {
    
    private static Commune.Builder commune;
    
    public CommuneTest() {
    }

    
     @Test
    public void communeTest1() {
        commune.communers(5).build();
        Assert.assertEquals(commune.build().getMaximumCommuners(), 5);
    }
    
     @Test
    public void communeTest2() {
        commune.communers(6).build();
        Assert.assertEquals(commune.build().getMaximumCommuners(), 6);
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
        ApplicationContext ctx = new AnnotationConfigApplicationContext(appConfiguration.class);
        commune = (Commune.Builder)ctx.getBean("commune");
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}
